requirements.txt
# Core
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Async
asyncio==3.4.3

# Web Scraping
playwright==1.40.0

# LLM
anthropic==0.40.0

# Embeddings
sentence-transformers==2.3.1

# Vector Store
chromadb==0.4.22

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Scheduler
apscheduler==3.10.4

# PDF Parsing
pymupdf==1.23.8
pdfplumber==0.10.3

# Data
numpy==1.26.3
pandas==2.1.4

# HTTP
httpx==0.26.0
aiohttp==3.9.1

# Utilities
tenacity==8.2.3
python-dateutil==2.8.2
beautifulsoup4==4.12.3
lxml==5.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
